added more solutions

+++ src/two-knights.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main() {
 	ll k;
 	cin >> k;
-	for(int i=1;i<=k;++i) {
+	for(ll i=1;i<=k;++i) {
 		cout << i * i * (i*i-1)/2 - 4*(i-1)*(i-2) <<endl;
 	}
 	return 0;
